homework06/hasker/requirements/base.txt
Django==1.11.4
django-admin-honeypot==1.0.0
djangorestframework==3.6.3
psycopg2==2.7.3
python-decouple==3.1
Pillow==4.2.1
